+++ changelog.md
@@ -1,5 +1,11 @@
 # Changelog
 
+# v0.2.0
+
+Added updateUser method
+Added deleteUser method
+Added testing
+
 # v0.1.0
 
 Initial SDK Setup